move times_result
